Update decorators.py
From the official python docs: 

get(key[, default])
Return the value for key if key is in the dictionary, else default. If default is not given, it defaults to None, so that this method never raises a KeyError.

Specifying a keyword argument causes an error, at least on my machine. 

Ubuntu 20.04.5
Python 3.9.16

https://docs.python.org/3.10/library/stdtypes.html?highlight=dict%20get#dict.get