feat: Support multiple additional clients token validation
Add a new config attribute called "ADDITIONAL_AWS_COGNITO_USER_POOL_CLIENT_IDS", this is a client ids list that enables the flask app to validate multiple clients of the same aws cognito user pool, this is important in a multi-client and single backend architectures, for instance mobile, web, and desktop applications using the same flask application as backend.
Testing user pool information was updated because a new access token and id token were needed in a different client of the same user pool of principal access token and id token of the tests, testing parameters can be restored to previous values if an additional access token and id token are created with another client of the previous user pool.